linter fixes and small structure refactoring
