+++ .golangci.yaml
@@ -31,23 +31,6 @@ linters-settings:
     range-loops: true
     # Report preallocation suggestions on for loops.
     for-loops: true
-  wsl:
-    # If true append is only allowed to be cuddled if appending value is matching variables, fields or types on line above
-    strict-append: true
-    # Allow calls and assignments to be cuddled as long as the lines have any matching variables, fields or types.
-    allow-assign-and-call: false
-    # Allow multiline assignments to be cuddled.
-    allow-multiline-assign: false
-    # Allow declarations (var) to be cuddled.
-    allow-cuddle-declarations: false
-    # Allow trailing comments in ending of blocks.
-    allow-trailing-comment: false
-    # Force newlines in end of case at this limit (0 = never).
-    force-case-trailing-whitespace: 0
-    # Force cuddling of err checks with err var assignment.
-    force-err-cuddling: true
-    # Allow leading comments to be separated with empty lines.
-    allow-separated-leading-comment: false
 
 linters:
   # please, do not use `enable-all`: it's deprecated and will be removed soon.
@@ -66,4 +49,3 @@ linters:
     - noctx # Doesn't have any config for linters-settings section.
     - prealloc
     - testpackage
-    - wsl
