+++ config/config.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	"github.com/caarlos0/env"
+	"github.com/caarlos0/env/v6"
 )
 
 type Config struct {
