+++ go.mod
@@ -3,7 +3,7 @@ module github.com/lungria/spendshelf-backend
 go 1.15
 
 require (
-	github.com/caarlos0/env/v6 v6.3.0 // indirect
+	github.com/caarlos0/env/v6 v6.3.0
 	github.com/jackc/pgx/v4 v4.8.1
 	github.com/rs/zerolog v1.20.0
 	github.com/stretchr/testify v1.6.1
