+++ main.go
@@ -10,14 +10,12 @@ import (
 
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/lungria/spendshelf-backend/config"
-	"github.com/lungria/spendshelf-backend/importer"
-	"github.com/lungria/spendshelf-backend/importer/interval"
 	"github.com/lungria/spendshelf-backend/job"
 	"github.com/lungria/spendshelf-backend/mono"
+	"github.com/lungria/spendshelf-backend/mono/importer"
+	"github.com/lungria/spendshelf-backend/mono/importer/interval"
 	"github.com/lungria/spendshelf-backend/storage"
 )
-
-const dbConnString = "postgres://localhost:5432/postgres?sslmode=disable"
 
 func main() {
 	cfg, err := config.FromEnv()
@@ -28,21 +26,24 @@ func main() {
 
 	bckgCtx := context.Background()
 	ctx, cancel := context.WithCancel(bckgCtx)
+
 	defer cancel()
 
-	apiClient := mono.NewClient(cfg.MonoBaseURL, cfg.MonoAPIKey)
 	dbpool, err := pgxpool.Connect(context.Background(), cfg.DBConnString)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Unable to connect to database: %v\n", err)
 		os.Exit(1)
 	}
+
 	defer dbpool.Close()
+
 	s := storage.NewPostgreSQLStorage(dbpool)
-	intervalGen := interval.NewSimpleIntervalGenerator(s)
+	intervalGen := interval.NewIntervalGenerator(s)
+	apiClient := mono.NewClient(cfg.MonoBaseURL, cfg.MonoAPIKey)
 	i := importer.NewImporeter(apiClient, s, intervalGen)
 
 	scheduler := job.Scheduler{}
-	scheduler.Schedule(ctx, i.Import(cfg.MonoAccountID), 1*time.Minute)
+	scheduler.Schedule(ctx, i.Import(cfg.MonoAccountID), 1*time.Minute, 30*time.Second)
 
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
