+++ mono/importer/interval/interval.go
@@ -1,0 +1,48 @@
+package interval
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// maxAllowedIntervalDuration limits max length of interval (in seconds) that mono API allows us to query.
+const maxAllowedIntervalDuration = 2682000
+
+// TransactionsStorage abstracts data access layer for already imported transactions.
+type TransactionsStorage interface {
+	// GetLastTransactionDate returns date property of latest transaction (sorted by date desc).
+	GetLastTransactionDate(ctx context.Context, accountID string) (time.Time, error)
+}
+
+// Generator creates interval based on latest stored transaction.
+type Generator struct {
+	storage TransactionsStorage
+}
+
+// NewIntervalGenerator creates new instance of Generator.
+func NewIntervalGenerator(storage TransactionsStorage) *Generator {
+	return &Generator{storage: storage}
+}
+
+// GetInterval creates interval based on latest stored transaction. It will return error if there latest transaction was
+// created more than maxAllowedIntervalDuration seconds ago.
+func (gen *Generator) GetInterval(ctx context.Context, accountID string) (from, to time.Time, err error) {
+	lastKnownTransactionDate, err := gen.storage.GetLastTransactionDate(ctx, accountID)
+	if err != nil {
+		return time.Time{}, time.Time{}, err
+	}
+
+	nowUtc := time.Now().UTC()
+
+	diffSecs := nowUtc.Sub(lastKnownTransactionDate.UTC()).Seconds()
+	if diffSecs > maxAllowedIntervalDuration {
+		return time.Time{}, time.Time{}, intervalTooLongErr(lastKnownTransactionDate)
+	}
+
+	return lastKnownTransactionDate.UTC(), nowUtc, nil
+}
+
+func intervalTooLongErr(lastKnownTransactionDate time.Time) error {
+	return fmt.Errorf("interval too long, lastKnownTransactionDate: %v", lastKnownTransactionDate)
+}
