+++ mono/mono.go
@@ -7,8 +7,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"strconv"
-	"time"
 )
 
 type Client struct {
@@ -58,6 +56,7 @@ func (c *Client) performRequest(ctx context.Context, url, method string, request
 
 		return nil, apiErr
 	}
+
 	return body, nil
 }
 
@@ -67,23 +66,4 @@ type APIError struct {
 
 func (err APIError) Error() string {
 	return err.Description
-}
-
-// Time defines a timestamp encoded as epoch seconds in JSON
-type Time time.Time
-
-// MarshalJSON is used to convert the timestamp to JSON
-func (t Time) MarshalJSON() ([]byte, error) {
-	return []byte(strconv.FormatInt(time.Time(t).Unix(), 10)), nil
-}
-
-// UnmarshalJSON is used to convert the timestamp from JSON
-func (t *Time) UnmarshalJSON(s []byte) (err error) {
-	r := string(s)
-	q, err := strconv.ParseInt(r, 10, 64)
-	if err != nil {
-		return err
-	}
-	*(*time.Time)(t) = time.Unix(q, 0)
-	return nil
 }
