+++ mono/transaction.go
@@ -5,9 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-
-	"github.com/lungria/spendshelf-backend/transaction"
-
 	"strconv"
 	"strings"
 	"time"
@@ -21,59 +18,43 @@ type GetTransactionsQuery struct {
 
 func (q *GetTransactionsQuery) asRoute() string {
 	var sb strings.Builder
+
 	sb.WriteString("/")
 	sb.WriteString(q.Account)
 	sb.WriteString("/")
 	sb.WriteString(strconv.FormatInt(q.From.Unix(), 10))
 	sb.WriteString("/")
 	sb.WriteString(strconv.FormatInt(q.To.Unix(), 10))
+
 	return sb.String()
 }
 
-func (c *Client) GetTransactions(ctx context.Context, query GetTransactionsQuery) ([]transaction.Transaction, error) {
+func (c *Client) GetTransactions(ctx context.Context, query GetTransactionsQuery) ([]Transaction, error) {
 	uri := fmt.Sprintf("%s/personal/statement%s", c.baseURL, query.asRoute())
 
 	response, err := c.performRequest(ctx, uri, http.MethodGet, nil)
 	if err != nil {
 		return nil, fmt.Errorf("unable to get transactions: %w", err)
 	}
+
 	if len(response) == 0 {
 		return nil, fmt.Errorf("unable to get transactions: empty response without error received")
 	}
 
-	transactions := make(responseTransactions, 0)
+	transactions := make([]Transaction, 0)
 	if err = json.Unmarshal(response, &transactions); err != nil {
 		return nil, fmt.Errorf("failed unmarshal transactions form json body: %w", err)
 	}
-	return transactions.AsPublicAPIModel(), nil
+
+	return transactions, nil
 }
 
 // todo come up with better naming
-type responseTransaction struct {
+type Transaction struct {
 	ID          string `json:"id"`
 	Time        Time   `json:"time"`
 	Description string `json:"description"`
 	MCC         int32  `json:"mcc"`
 	Hold        bool   `json:"hold"`
 	Amount      int64  `json:"amount"`
-}
-
-// todo come up with better naming
-type responseTransactions []responseTransaction
-
-// todo come up with better naming
-func (t responseTransactions) AsPublicAPIModel() []transaction.Transaction {
-	transactions := make([]transaction.Transaction, len(t))
-	for i, v := range t {
-		transactions[i] = transaction.Transaction{
-			BankID:      v.ID,
-			Time:        time.Time(v.Time),
-			Description: v.Description,
-			MCC:         v.MCC,
-			Hold:        v.Hold,
-			Amount:      v.Amount,
-		}
-	}
-
-	return transactions
 }
